implement investment cards: add assets, conditional rendering and mobile view
